# php-tutor/php-learn/app/cache/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "php-learn API",
        "description": "PHP 8.5 无框架后端 + Vue 3 SPA 前后端分离项目的 REST API。\n\n鉴权方式：Bearer Token（JWT Access + Refresh Token）\n响应结构：统一 envelope {code, message, data, errors?}\n业务 code：0=成功, 40101=未登录, 40102=token 无效, 40301=无权限, 40401=资源不存在, 42201=校验失败, 50000=内部错误",
        "contact": {
            "name": "php-learn",
            "url": "https:\/\/github.com\/example\/php-learn"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "\/api\/v1",
            "description": "API v1（同域部署）"
        },
        {
            "url": "http:\/\/127.0.0.1:8000\/api\/v1",
            "description": "本地开发"
        }
    ],
    "paths": {
        "\/activities": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "活动列表",
                "description": "支持分页，upcoming=true 只返回即将开始的活动",
                "operationId": "a8170488c07f400aac287f1c6c1ff895",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页码",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "每页条数",
                        "schema": {
                            "type": "integer",
                            "default": 8,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "upcoming",
                        "in": "query",
                        "description": "传 true 时只返回未来的活动",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "活动分页列表"
                    }
                }
            },
            "post": {
                "tags": [
                    "Activities"
                ],
                "summary": "新建活动",
                "description": "需要 Bearer Token",
                "operationId": "4c1a7613c6d25f9b0bd5aa280c4e1c9b",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/Activity"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "参数校验失败",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "\/activities\/{id}": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "活动详情",
                "description": "获取指定 ID 的活动详情",
                "operationId": "22d2fceeee1395777adb5e35cb4b0d53",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "活动 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "活动详情"
                    },
                    "404": {
                        "description": "活动不存在"
                    }
                }
            },
            "put": {
                "tags": [
                    "Activities"
                ],
                "summary": "更新活动",
                "description": "PATCH 语义：只更新提交了字段",
                "operationId": "f68d2a5a85eb2f5111c90c0dc51236fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "活动 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/Activity"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "活动不存在",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "参数校验失败",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Activities"
                ],
                "summary": "删除活动",
                "description": "DELETE \/api\/v1\/activities\/{id}（需鉴权）",
                "operationId": "a7e52a4c35d0fa9cbaadd7f61c6fd498",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "活动 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "删除成功（无 body）"
                    },
                    "401": {
                        "description": "未登录",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "活动不存在",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer Token（JWT Access Token），通过登录接口获取",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}